requirements.txt
tensorflow
numpy
opencv-python
scikit-learn
matplotlib
pillow
pillow-heif>=0.16.0
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
supabase
httpx
pytest
pytest-asyncio
jupyter
tqdm
python-dotenv
